Split check file-read host effect fixtures
