+++ tests/integration/cli_build/graph_validation_host_effects/file_reads/mod.rs
@@ -1,0 +1,2 @@
+mod declared;
+mod rejections;
