+++ tests/integration/cli_build/graph_validation_host_effects/file_reads/rejections.rs
@@ -1,166 +1,6 @@
 use std::process::Command;
 
 #[test]
-fn check_command_build_zen_accepts_declared_file_read_effects() {
-    assert_check_command_accepts_declared_file_read_effect(
-        r#"| .Err { "default" }"#,
-        "check_command_build_zen_accepts_declared_file_read_effects",
-    );
-}
-
-#[test]
-fn check_command_build_zen_accepts_wildcard_fallback_declared_file_read_effects() {
-    assert_check_command_accepts_declared_file_read_effect(
-        r#"| _ { "default" }"#,
-        "check_command_build_zen_accepts_wildcard_fallback_declared_file_read_effects",
-    );
-}
-
-#[test]
-fn check_command_build_zen_accepts_identifier_fallback_declared_file_read_effects() {
-    assert_check_command_accepts_declared_file_read_effect(
-        r#"| err { "default" }"#,
-        "check_command_build_zen_accepts_identifier_fallback_declared_file_read_effects",
-    );
-}
-
-fn assert_check_command_accepts_declared_file_read_effect(fallback_arm: &str, case_name: &str) {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    std::fs::write(
-        tmp.path().join("build.zen"),
-        format!(
-            r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {{
-    manifest = b.os.read_file("build.targets") ?
-        | .Ok(contents) {{ contents }}
-        {fallback_arm}
-    b.add(Executable {{ name: "myapp", main: "main.zen", out_dir: "build/" }})
-    .Ok(b.config())
-}}
-"#,
-        ),
-    )
-    .expect("write build.zen");
-    std::fs::write(tmp.path().join("build.targets"), "myapp\n").expect("write manifest");
-    std::fs::write(
-        tmp.path().join("main.zen"),
-        r#"
-main = () i32 {
-    0
-}
-"#,
-    )
-    .expect("write main.zen");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .args(["check", "build.zen"])
-        .current_dir(tmp.path())
-        .output()
-        .expect("run zen check build.zen");
-
-    assert!(
-        output.status.success(),
-        "{case_name}: zen check build.zen failed: stdout={}, stderr={}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-    assert!(
-        String::from_utf8_lossy(&output.stdout).contains("1 build targets"),
-        "expected build graph check summary, stdout={}",
-        String::from_utf8_lossy(&output.stdout)
-    );
-}
-
-#[test]
-fn check_command_build_zen_accepts_declared_file_read_effects_for_multiple_targets() {
-    assert_check_command_accepts_declared_file_read_effects_for_multiple_targets(
-        r#"| .Err { "default" }"#,
-        "check_command_build_zen_accepts_declared_file_read_effects_for_multiple_targets",
-    );
-}
-
-#[test]
-fn check_command_build_zen_accepts_wildcard_fallback_declared_file_read_effects_for_multiple_targets(
-) {
-    assert_check_command_accepts_declared_file_read_effects_for_multiple_targets(
-        r#"| _ { "default" }"#,
-        "check_command_build_zen_accepts_wildcard_fallback_declared_file_read_effects_for_multiple_targets",
-    );
-}
-
-#[test]
-fn check_command_build_zen_accepts_identifier_fallback_declared_file_read_effects_for_multiple_targets(
-) {
-    assert_check_command_accepts_declared_file_read_effects_for_multiple_targets(
-        r#"| err { "default" }"#,
-        "check_command_build_zen_accepts_identifier_fallback_declared_file_read_effects_for_multiple_targets",
-    );
-}
-
-fn assert_check_command_accepts_declared_file_read_effects_for_multiple_targets(
-    fallback_arm: &str,
-    case_name: &str,
-) {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    std::fs::write(
-        tmp.path().join("build.zen"),
-        format!(
-            r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {{
-    manifest = b.os.read_file("build.targets") ?
-        | .Ok(contents) {{ contents }}
-        {fallback_arm}
-    b.add(Executable {{ name: "app", main: "app.zen", out_dir: "build/app/" }})
-    b.add(Test {{ name: "unit", root: "unit.zen" }})
-    b.add(Library {{ name: "core", exports: ["lib.zen"] }})
-    .Ok(b.config())
-}}
-"#,
-        ),
-    )
-    .expect("write build.zen");
-    std::fs::write(tmp.path().join("build.targets"), "app\nunit\ncore\n").expect("write manifest");
-    for source in ["app.zen", "unit.zen"] {
-        std::fs::write(
-            tmp.path().join(source),
-            r#"
-main = () i32 {
-    0
-}
-"#,
-        )
-        .unwrap_or_else(|err| panic!("write {source}: {err}"));
-    }
-    std::fs::write(
-        tmp.path().join("lib.zen"),
-        r#"
-value = () i32 {
-    1
-}
-"#,
-    )
-    .expect("write lib.zen");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .args(["check", "build.zen"])
-        .current_dir(tmp.path())
-        .output()
-        .expect("run zen check build.zen");
-
-    assert!(
-        output.status.success(),
-        "{case_name}: zen check build.zen failed: stdout={}, stderr={}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-    assert!(
-        String::from_utf8_lossy(&output.stdout).contains("3 build targets"),
-        "expected multi-target build graph check summary, stdout={}",
-        String::from_utf8_lossy(&output.stdout)
-    );
-}
-
-#[test]
 fn check_command_build_zen_rejects_undeclared_file_read_effects_before_source_validation() {
     let tmp = tempfile::tempdir().expect("create temp dir");
     std::fs::write(
